feat(노션 편집기) : 노션 편집기 1차 구현 완료

+++ src/pages/MainPage.js
@@ -1,12 +1,15 @@
 import Editor from "../components/Editor.js";
 import Sidebar from "../components/Sidebar.js";
 import { request } from "../utils/api.js";
+import { push, replace } from "../utils/router.js";
 import { getLocalStorageItem, removeLocalStorageItem, setLocalStorageItem } from "../utils/storage.js";
 
 export default function MainPage({$target, initialState}){
   const $page = document.createElement("div");
   this.state = initialState;
-  let postLocalSaveKey = `temp-post-${this.state.id}`;
+  const getPostLocalSaveKey = (id) =>{
+    return `temp-post-${id}`;
+  } 
   
   const fetchDocuments = async () => {
     const documents = await request("/documents");
@@ -17,30 +20,50 @@ export default function MainPage({$target, initialState}){
   }
 
   const fetchContent = async (id) => {
-    const document = await request(`/documents/${id}`);
+    const documentContent = await request(`/documents/${id}`);
+    const tempContent = getLocalStorageItem(getPostLocalSaveKey(id), null);
+    if(tempContent && tempContent.tempSaveDate > documentContent.updatedAt){
+      if(confirm("저장된 작성글이 있습니다. 불러올까요?")){    
+        this.setState({
+          ...this.state,
+          id,
+          tempContent
+        })
+        return;
+      }
+    }
     this.setState({
       ...this.state,
       id,
-      document
+      documentContent
     })
   }
 
   this.setState = (nextState) => {
-    if(this.state.id !== nextState.id)
-      editor.setState(nextState.document);
+    const currentId = this.state.id;
     this.state = nextState;
-    sideBar.setState(this.state.documents);
+    sideBar.setState({
+      documents : this.state.documents, 
+      selectedId : this.state.id
+    });
+    if(currentId !== nextState.id){
+      editor.setState(this.state.documentContent);
+    }
     this.render();
   };
-
+  
   this.render =  () => {
     $target.appendChild($page);
   }
+  this.render();
 
   let timer = null;
   const sideBar = new Sidebar({
     $target : $page,
-    initialState : this.state.documents,
+    initialState : {
+      selectedId : this.state.id,
+      documents : this.state.documents
+    },
     onClickAdd : async (id) => {
       const newDocument = {
         title : "",
@@ -49,49 +72,67 @@ export default function MainPage({$target, initialState}){
       newDocument.title = "빈 제목";
       newDocument.parent = id;
       
-      await request("/documents", {
+      const {id : newId} = await request("/documents", {
         method : "POST",
         body : JSON.stringify(newDocument)
       })
+      push(`/documents/${newId}`);
       await fetchDocuments();
+      await fetchContent(newId);
     },
     onClickDelete : async (id) => {
-      await request(`/documents/${id}`, {
-        method : "DELETE"
-      })
-      fetchDocuments();
+      if(confirm("정말로 삭제하시겠습니까?")){
+        await request(`/documents/${id}`, {
+          method : "DELETE"
+        })
+        replace("/");
+        fetchDocuments();
+        this.setState({
+          ...this.state,
+          documentContent : null
+        })
+      }
     },
     onClickDocument : async (id) => {
+      push(`/documents/${id}`);
       await fetchContent(id);
     }
   });
 
   const editor = new Editor({
     $target : $page, 
-    initialState : {
-      title : "",
-      content : ""
-    }, 
-    onEditing : (post, $input) => {
+    initialState : this.state.documentContent, 
+    onEditing : (post, id, $input) => {
       if(timer !== null){
         clearTimeout(timer);
       }
       timer = setTimeout(async () => {
-        setLocalStorageItem(`temp-post-${this.state.id}`, {
+        setLocalStorageItem(getPostLocalSaveKey(id), {
           ...post,
           tempSaveDate : new Date()
         })
-        await request(`/documents/${this.state.id}`, {
+        await request(`/documents/${id}`, {
           method : "PUT",
           body : JSON.stringify(post)
         })
         await fetchDocuments();
-        removeLocalStorageItem(postLocalSaveKey);
+        removeLocalStorageItem(getPostLocalSaveKey(id));
         $input.focus();
       }, 500)
     }});
-  this.init = async () => {
-    await fetchDocuments();
+
+  this.init = async (id) => {
     this.render();
+    await fetchDocuments();
+    if(id){
+      await fetchContent(id);
+    }else{
+      this.setState({
+        ...this.state,
+        id : null,
+        documentContent : null
+      })
+    }
   }
+
 }
